==> R/16_campus_wifi_green_bridge.Rmd <==
---
title: "POP247 - campus EDA"
subtitle: "Models for St Lucia - Wi-Fi & Green Bridge (pedestrians + cyclists)"
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      fig.width=8, fig.height=6, dpi=300, 
                      out.width="800px", out.height="600px",
                      cache=FALSE)

set.seed(12345)
options(scipen = 999)

library(pacman) 
p_load(tidyverse, magrittr, lubridate, corrplot, scales, sugrrants,
       fitdistrplus, INLA, INLAutils)

# inla.setOption(num.threads = 1) 
```

```{r conflicts, include = FALSE}
tidyverse::tidyverse_conflicts()
conflicted::conflict_scout()
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)
knitr::opts_knit$set(width = 75)
```

<!-- ------------------------------------------------------------ --> 

# Pedestrian & bike counts

## Original data

```{r echo=FALSE}
green_bridge <- readRDS("data/counters/clean/green_bridge.Rds") %>% 
  dplyr::select(-site) %>% 
  filter(date >= ymd_hms("2018-12-03 00:00:00", tz = "Australia/Brisbane")) %>% 
  filter(date <= ymd_hms("2019-01-27 23:59:59", tz = "Australia/Brisbane")) 
```

Detailed example counts for week of `2018-12-03` across type and direction.

```{r echo=FALSE}
green_bridge %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  mutate(count = ifelse(test = direction == "outbound", yes = -1 * count, no = count)) %>%
  ggplot(aes(x = date, y = count)) + 
  geom_line(aes(group = interaction(type, direction), color = type)) + 
  scale_x_datetime() + theme_minimal() + 
  xlab("") + ylab("Bike / pedestrian counts")
```

Overview of months, cyclists, outbound

```{r echo=FALSE, warning=FALSE}
p <- green_bridge %>%
  filter(direction == "outbound") %>%
  filter(type == "cyclists") %>%
  rename(date_time = date) %>% 
  mutate(weekday = strftime(date_time, format = "%A", tz = "Australia/Brisbane"),
         weekend = if_else(weekday %in% c("Saturday", "Sunday"), "Weekend", "Weekday"),
         date = as.Date(date_time, tz = "Australia/Brisbane"),
         time = as.numeric(format(date_time, "%H")) + as.numeric(format(date_time, "%M"))/60) %>%
  frame_calendar(x = time, y = count, date = date) %>% 
  ggplot(aes(x = .time, y = .count, group = date, colour = weekend)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal()

prettify(p)
```

## Stocks on campus 

### Without modification

```{r echo=FALSE}
green_bridge_wide <- green_bridge %>% 
  mutate(count = ifelse(test = direction == "outbound", yes = -1 * count, no = count)) %>%
  unite(temp, type, direction) %>%
  spread(temp, count) %>% 
  mutate(net = cyclists_inbound + cyclists_outbound + pedestrians_inbound + pedestrians_outbound) %>% 
  mutate(stock = cumsum(net)) 

green_bridge_wide %>% 
  ggplot(aes(x = date, y = stock)) + 
  geom_line() + 
  scale_x_datetime() + theme_minimal() + 
  xlab("") + ylab("Stocks from bike & pedestrian counts")
```

### With midnight 'reset'

```{r echo=FALSE}
green_bridge_wide <- green_bridge %>% 
  mutate(count = ifelse(test = direction == "outbound", yes = -1 * count, no = count)) %>%
  unite(temp, type, direction) %>%
  spread(temp, count) %>% 
  mutate(net = cyclists_inbound + cyclists_outbound + pedestrians_inbound + pedestrians_outbound) %>% 
  group_by(date(date)) %>%
  mutate(stock = cumsum(net)) %>% 
  ungroup()

green_bridge_wide %>% 
  ggplot(aes(x = date, y = stock)) + 
  geom_line() + 
  scale_x_datetime() + theme_minimal() + 
  xlab("") + ylab("Stocks from bike & pedestrian counts")
```

Frequency of net from whole period

```{r echo=FALSE}
green_bridge_wide %>% 
  ggplot(aes(x = net)) + 
  geom_histogram(binwidth = 1) + 
  theme_minimal() + 
  xlab("") + ylab("Wi-Fi users") 
```

Frequency of stocks from whole period

```{r echo=FALSE}
green_bridge_wide %>% 
  ggplot(aes(x = stock)) + 
  geom_histogram(binwidth = 10) + 
  theme_minimal() + 
  xlab("") + ylab("Wi-Fi users") 
```

<!-- ------------------------------------------------------------ --> 

# Wi-Fi

## Original data 

```{r echo=FALSE}
prime_agg <- readRDS("../TEMPO_Wi-Fi/data/clean/processed_prime_reports/prime_agg_3.Rds") %>% 
  rename(date = prime_agg_3) %>% 
  filter(date >= ymd_hms("2018-12-03 00:00:00", tz = "Australia/Brisbane")) %>% 
  filter(date <= ymd_hms("2019-01-27 23:59:59", tz = "Australia/Brisbane")) 
  # filter(date <= ymd_hms("2019-03-25 23:59:59", tz = "Australia/Brisbane"))

# tz(prime_agg$date)
```

Detailed example counts for week of `2018-12-09` across type of data.

```{r echo=FALSE}
prime_agg %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  ggplot(aes(x = date, y = users_mac)) + 
  geom_line() +
  scale_x_datetime() + theme_minimal() + 
  xlab("") + ylab("Wi-Fi users")

```

Overview of months, connected users

```{r echo=FALSE, warning=FALSE}
p <- prime_agg %>%
  mutate(weekday = strftime(date, format = "%A", tz = "Australia/Brisbane"),
         weekend = if_else(weekday %in% c("Saturday", "Sunday"), "Weekend", "Weekday"),
         time = as.numeric(format(date, "%H", tz = "Australia/Brisbane")) + 
           as.numeric(format(date, "%M", tz = "Australia/Brisbane"))/60,
         date2 = as.Date(date, tz = "Australia/Brisbane")) %>%
  frame_calendar(x = time, y = users_mac, date = date2) %>% 
  ggplot(aes(x = .time, y = .users_mac, group = date2, colour = weekend)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal()

prettify(p)
```


## Distribution of counts

Frequency of counts from whole period

```{r echo=FALSE}
prime_agg %>% 
  ggplot(aes(x = users_mac)) + 
  geom_histogram(binwidth = 100) + 
  theme_minimal() + 
  xlab("") + ylab("Wi-Fi users") 
```

## Distribution fits 

```{r echo=FALSE}
prime_agg_poi <- fitdist(prime_agg$users_mac, "pois")
prime_agg_nb <- fitdist(prime_agg$users_mac, "nbinom")

par(mfrow = c(1,2))
denscomp(list(prime_agg_poi, prime_agg_nb), legendtext = c("Poisson", "negative binomial"), fitlty = 1)
cdfcomp(list(prime_agg_poi, prime_agg_nb), legendtext = c("Poisson", "negative binomial"), fitlty = 1)

# gofstat(list(prime_agg_poi, prime_agg_nb), fitnames = c("Poisson", "negative binomial"))
```

```{r message=FALSE, warning=FALSE, include=FALSE}
dev.off()
rm(prime_agg_poi, prime_agg_nb)
```

<!-- ------------------------------------------------------------ --> 

# Time series comparisons

```{r echo=FALSE}
data <- full_join(prime_agg, green_bridge_wide) %>% 
  filter(!is.na(users_mac)) %>% 
  mutate(weekday = strftime(date, format = "%A", tz = "Australia/Brisbane"),
         weekend = if_else(weekday %in% c("Saturday", "Sunday"), "Weekend", "Weekday"),
         time = as.numeric(date - min(prime_agg$date)) / 900) %>% 
  dplyr::select(-`date(date)`)

rm(prime_agg, green_bridge_wide)

# data_long <- data %>% 
#   dplyr::select(date, time, weekday, weekend, users_mac, 
#          cyclists_inbound:pedestrians_outbound) %>% 
#   gather(key = "site_type", value = "count", cyclists_inbound:pedestrians_outbound)
# 
# data_long$time2 <- data_long$time
# data_long$time3 <- data_long$time
# data_long$time4 <- data_long$time
# 
# data_long$cyclists_inbound <- as.numeric(data_long$site_type == "cyclists_inbound")
# data_long$cyclists_outbound <- as.numeric(data_long$site_type == "cyclists_outbound")
# data_long$pedestrians_inbound <- as.numeric(data_long$site_type == "pedestrians_inbound")
# data_long$pedestrians_outbound <- as.numeric(data_long$site_type == "pedestrians_outbound")
```

## Cyclists 

### Inbound, connected users

```{r echo=FALSE}
data %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  dplyr::select(date, users_mac, cyclists_inbound) %>% 
  gather(key = "type", value = "count", users_mac:cyclists_inbound) %>% 
  ggplot(aes(x = date, y = count)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal() + ylab("Connected users") +
  facet_wrap(~type, ncol = 1, scales = "free_y")
```

### Outbound, connected users

```{r echo=FALSE}
data %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  dplyr::select(date, users_mac, cyclists_outbound) %>% 
  gather(key = "type", value = "count", users_mac:cyclists_outbound) %>% 
  ggplot(aes(x = date, y = count)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal() + ylab("Connected users") +
  facet_wrap(~type, ncol = 1, scales = "free_y")
```

## Pedestrians 

### Inbound, connected users

```{r echo=FALSE}
data %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  dplyr::select(date, users_mac, pedestrians_inbound) %>% 
  gather(key = "type", value = "count", users_mac:pedestrians_inbound) %>% 
  ggplot(aes(x = date, y = count)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal() + ylab("Connected users") +
  facet_wrap(~type, ncol = 1, scales = "free_y")
```

### Outbound, connected users

```{r echo=FALSE}
data %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  dplyr::select(date, users_mac, pedestrians_outbound) %>% 
  gather(key = "type", value = "count", users_mac:pedestrians_outbound) %>% 
  ggplot(aes(x = date, y = count)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal() + ylab("Connected users") +
  facet_wrap(~type, ncol = 1, scales = "free_y")
```

## Stocks, connected users

```{r echo=FALSE}
data %>% 
  filter(date <= ymd_hms("2018-12-09 23:59:59", tz = "Australia/Brisbane")) %>%
  dplyr::select(date, users_mac, stock) %>% 
  gather(key = "type", value = "count", users_mac:stock) %>% 
  ggplot(aes(x = date, y = count)) +
  geom_line() +
  theme(legend.position = "none") + 
  theme_minimal() + ylab("Connected users") +
  facet_wrap(~type, ncol = 1, scales = "free_y")
```

# Correlations, connected users 

```{r echo=FALSE}
M <- cor(dplyr::select(data, users_mac, 
                       cyclists_inbound, cyclists_outbound, 
                       pedestrians_inbound, pedestrians_outbound, 
                       net, stock), use = "complete.obs")

corrplot.mixed(M)
```

<!-- ------------------------------------------------------------ --> 
# Models

## INLA setup

```{r echo=TRUE, results=FALSE}
# ####
# Jeffreys prior 
a1 <- 5e-5
b1 <- 5e-5
lgprior1 <- list(prec = list(param = c(a1, b1)))

# Gelman prior
a2 <- -0.5
b2 <- 5e-5
lgprior2 <- list(prec = list(param = c(a2, b2)))

# iid prior 
# Schrödle & Held 2010 & Blangiardo et al 2013
a0 <- 1
b0 <- 0.1
prior.nu <- list(prec = list(param = c(a0, b0)))

# intercept & fixed
inla.set.control.fixed.default() 

# intercept ~ N(0,0) 
# other fixed effects ~ N(0, 0.001) 
# 
# where the format is N(mean, precision) 
# precision = inverse of the variance. 

# PC prior
U <- 1
hyper.prec = list(theta = list(
  prior = "pc.prec",
  param = c(U, 0.01)
))

# scaling 
inla.setOption(scale.model.default = TRUE)

```

## RW1

### One trend for stocks

```{r}
mod1 <- inla(users_mac ~ 
               f(stock, model = "rw1", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod1)
autoplot(mod1)

P1 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod1$summary.fitted.values$`0.5quant`,
                lci = mod1$summary.fitted.values$`0.025quant`, 
                uci = mod1$summary.fitted.values$`0.975quant`)

ggplot(P1, aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P1 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

### One trend for stocks + weekday

```{r}
mod2 <- inla(users_mac ~ 
               weekday +
               f(stock, model = "rw1", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod2)
autoplot(mod2)

P2 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod2$summary.fitted.values$`0.5quant`,
                lci = mod2$summary.fitted.values$`0.025quant`, 
                uci = mod2$summary.fitted.values$`0.975quant`)

ggplot(P2, aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P2 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

### One trend for stocks + weekend + weekday

```{r}
mod3 <- inla(users_mac ~ 
               weekend +
               weekday +
               f(stock, model = "rw1", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod3)
autoplot(mod3)

P3 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod3$summary.fitted.values$`0.5quant`,
                lci = mod3$summary.fitted.values$`0.025quant`, 
                uci = mod3$summary.fitted.values$`0.975quant`)

P3 %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P3 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

## RW2

### One trend for stocks

```{r}
mod5 <- inla(users_mac ~ 
               # weekend +
               # weekday +
               f(stock, model = "rw2", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod5)
autoplot(mod5)

P5 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod5$summary.fitted.values$`0.5quant`,
                lci = mod5$summary.fitted.values$`0.025quant`, 
                uci = mod5$summary.fitted.values$`0.975quant`)

ggplot(P5, aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P5 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

### One trend for stocks + weekday

```{r}
mod6 <- inla(users_mac ~ 
               # weekend +
               weekday +
               f(stock, model = "rw2", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod6)
autoplot(mod6)

P6 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod6$summary.fitted.values$`0.5quant`,
                lci = mod6$summary.fitted.values$`0.025quant`, 
                uci = mod6$summary.fitted.values$`0.975quant`)

ggplot(P6, aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P6 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

### One trend for stocks + weekend + weekday

```{r}
mod7 <- inla(users_mac ~ 
               weekend +
               weekday +
               f(stock, model = "rw2", scale.model = TRUE, hyper = hyper.prec), 
             family = "nbinomial", 
             control.compute = list(dic= TRUE, waic = TRUE),  
             data = data)
```

```{r echo=FALSE}
summary(mod7)
autoplot(mod7)

P7 <- bind_cols(time = data$time, count = data$users_mac,
                fit = mod7$summary.fitted.values$`0.5quant`,
                lci = mod7$summary.fitted.values$`0.025quant`, 
                uci = mod7$summary.fitted.values$`0.975quant`)

P7 %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 

P7 %>% 
  filter(time >= 4704) %>% 
  ggplot( aes(x = time)) +
  geom_line(aes(y = count), colour = "orange", alpha = 0.5) +
  geom_line(aes(y = fit), col = "purple") +
  geom_ribbon(aes(ymin = lci, ymax = uci), alpha = 0.3) +
  theme_minimal() + xlab("") 
```

## Comparing

### DIC

```{r echo=FALSE}
dotchart(c(mod1$dic$dic, mod2$dic$dic, mod3$dic$dic,
           mod5$dic$dic, mod6$dic$dic, mod7$dic$dic),
         labels = c("mod1", "mod2", "mod3", "mod5", "mod6", "mod7"), main = "DIC")
```

### WAIC 

```{r echo=FALSE}
dotchart(c(mod1$waic$waic, mod2$waic$waic, mod3$waic$waic,
           mod5$waic$waic, mod6$waic$waic, mod7$waic$waic),
         labels = c("mod1", "mod2", "mod3", "mod5", "mod6", "mod7"), main = "WAIC")
```
